+++ exonum/src/blockchain/tests.rs
@@ -160,10 +160,7 @@ impl Transaction {
         public_key: PublicKey,
         secret_key: &SecretKey,
     ) -> Verified<AnyTx> {
-        let tx = AnyTx {
-            arguments: self.into_bytes(),
-            call_info: CallInfo::new(instance_id, 0),
-        };
+        let tx = AnyTx::new(CallInfo::new(instance_id, 0), self.into_bytes());
         Verified::from_value(tx, public_key, secret_key)
     }
 }
@@ -248,7 +245,7 @@ impl RuntimeInspector {
     }
 
     fn default_artifact_id() -> ArtifactId {
-        ArtifactId::new(Self::ID, "runtime-inspector", Version::new(1, 0, 0)).unwrap()
+        ArtifactId::from_raw_parts(Self::ID, "runtime-inspector".into(), Version::new(1, 0, 0))
     }
 }
 
@@ -519,8 +516,11 @@ fn initialize_service_ok() {
 fn deploy_available() {
     let (pk, sk) = exonum_crypto::gen_keypair();
 
-    let artifact_id =
-        ArtifactId::new(RuntimeInspector::ID, "secondary", Version::new(1, 0, 0)).unwrap();
+    let artifact_id = ArtifactId::from_raw_parts(
+        RuntimeInspector::ID,
+        "secondary".into(),
+        Version::new(1, 0, 0),
+    );
     let mut blockchain = create_blockchain(
         RuntimeInspector::default().with_available_artifact(artifact_id.clone()),
         vec![InitAction::Noop.into_default_instance()],
@@ -569,8 +569,11 @@ fn deploy_unavailable_artifact() {
         vec![InitAction::Noop.into_default_instance()],
     );
 
-    let artifact_id =
-        ArtifactId::new(RuntimeInspector::ID, "secondary", Version::new(1, 0, 0)).unwrap();
+    let artifact_id = ArtifactId::from_raw_parts(
+        RuntimeInspector::ID,
+        "secondary".into(),
+        Version::new(1, 0, 0),
+    );
     execute_transaction(
         &mut blockchain,
         Transaction::DeployArtifact(artifact_id).sign(TEST_SERVICE_ID, pk, &sk),
@@ -588,11 +591,11 @@ fn start_stop_service_instance() {
     );
 
     // Start secondary service instance.
-    let instance_spec = InstanceSpec {
-        id: 10,
-        name: "secondary".to_owned(),
-        artifact: RuntimeInspector::default_artifact_id(),
-    };
+    let instance_spec = InstanceSpec::from_raw_parts(
+        10,
+        "secondary".to_owned(),
+        RuntimeInspector::default_artifact_id(),
+    );
 
     // Check that the secondary service instance is absent in the dispatcher schema.
     let snapshot = blockchain.snapshot();