+++ exonum/src/runtime/mod.rs
@@ -210,6 +210,9 @@ mod types;
 pub const SUPERVISOR_INSTANCE_ID: InstanceId = 0;
 
 /// List of predefined runtimes.
+///
+/// This type is not intended to be exhaustively matched. It can be extended in the future
+/// without breaking the semver compatibility.
 #[derive(Debug, PartialEq, Eq, Hash, Clone)]
 #[repr(u32)]
 pub enum RuntimeIdentifier {
@@ -217,6 +220,10 @@ pub enum RuntimeIdentifier {
     Rust = 0,
     /// Exonum Java Binding runtime.
     Java = 1,
+
+    /// Never actually generated.
+    #[doc(hidden)]
+    __NonExhaustive,
 }
 
 impl From<RuntimeIdentifier> for u32 {
@@ -240,6 +247,7 @@ impl fmt::Display for RuntimeIdentifier {
         match self {
             RuntimeIdentifier::Rust => f.write_str("Rust runtime"),
             RuntimeIdentifier::Java => f.write_str("Java runtime"),
+            RuntimeIdentifier::__NonExhaustive => unreachable!("Never actually generated"),
         }
     }
 }
@@ -613,14 +621,25 @@ pub struct RuntimeInstance {
     pub id: u32,
     /// Enclosed `Runtime` object.
     pub instance: Box<dyn Runtime>,
+
+    /// No-op field for forward compatibility.
+    non_exhaustive: (),
+}
+
+impl RuntimeInstance {
+    /// Constructs a new `RuntimeInstance` object.
+    pub fn new(id: u32, instance: Box<dyn Runtime>) -> Self {
+        Self {
+            id,
+            instance,
+            non_exhaustive: (),
+        }
+    }
 }
 
 impl<T: WellKnownRuntime> From<T> for RuntimeInstance {
     fn from(runtime: T) -> Self {
-        RuntimeInstance {
-            id: T::ID,
-            instance: runtime.into(),
-        }
+        RuntimeInstance::new(T::ID, runtime.into())
     }
 }
 
@@ -789,6 +808,20 @@ pub struct InstanceDescriptor<'a> {
     /// A unique name of the service instance.
     /// [Read more.](struct.InstanceSpec.html#structfield.name)
     pub name: &'a str,
+
+    /// No-op field for forward compatibility.
+    non_exhaustive: (),
+}
+
+impl<'a> InstanceDescriptor<'a> {
+    /// Creates a new `InstanceDescriptor` object.
+    pub fn new(id: InstanceId, name: &'a str) -> Self {
+        Self {
+            id,
+            name,
+            non_exhaustive: (),
+        }
+    }
 }
 
 impl fmt::Display for InstanceDescriptor<'_> {
@@ -805,7 +838,7 @@ impl From<InstanceDescriptor<'_>> for (InstanceId, String) {
 
 impl<'a> From<(InstanceId, &'a str)> for InstanceDescriptor<'a> {
     fn from((id, name): (InstanceId, &'a str)) -> Self {
-        InstanceDescriptor { id, name }
+        InstanceDescriptor::new(id, name)
     }
 }